Update Character interface
The Character enum no longer represents the characters are isize,
instead it provides a method called as_char that converts the character
enum into a proper char. It also has a new Custom variant, where someone
can store a custom character, implements Into<char> and From<char> (the
first calls as_char and the second stores the char in a Custom variant)
and Display, which simply displays the character for the variant
selected.