New MaybeAsF64 trait
Creates and implements a new trait: MaybeAsF64. This trait is created to
allow to draw non continuous functions. It works basically equal to
AsF64, but instead of returning and f64 by calling as_f64, it returns
and Option<f64> by calling maybe_as_f64 (it does not have a
maybe_from_f64() as it wouldn't be useful). In a future, the return
type for Function will implement MaybeAsF64 instead of AsF64 to support
a wider range of functions.