Reorganize exports
The base crate now exports all public entities instead of manually
adding every one with the use of the asterisk wildcard.
A doc comment for the reexporting of console_engine::Color is added to
clarify its origin.
The original draw method is now private (and still not shown in
documentation).